+++ src/Test/UnitTestCase.php
@@ -4,15 +4,13 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Test;
 
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
-use Doctrine\DBAL\Query\Expression\ExpressionBuilder;
+use MakinaCorpus\QueryBuilder\Bridge\Mock\MockQueryBuilder;
+use MakinaCorpus\QueryBuilder\DatabaseSession;
 use MakinaCorpus\QueryBuilder\DefaultQueryBuilder;
 use MakinaCorpus\QueryBuilder\Expression;
+use MakinaCorpus\QueryBuilder\Platform\Escaper\StandardEscaper;
 use MakinaCorpus\QueryBuilder\QueryBuilder;
 use MakinaCorpus\QueryBuilder\SqlString;
-use MakinaCorpus\QueryBuilder\Platform\Escaper\StandardEscaper;
 use MakinaCorpus\QueryBuilder\Writer\Writer;
 use PHPUnit\Framework\TestCase;
 
@@ -20,9 +18,8 @@ use PHPUnit\Framework\TestCase;
  * Extends this class whenever you need either the connection or the query
  * builder from doctrine/dbal.
  *
- * Per default, it will yield the PostgreSQL database platform, it is meant
- * for writing unit test or anonymizers. If you need another dialect for a
- * certain test class, override createPlatform() method to change it.
+ * Per default, it will yield the PostgreSQL database platform, and it is
+ * hardcoded.
  *
  * If you need to test all dialects, either:
  *   - write unit test of every helper method you added,
@@ -30,77 +27,12 @@ use PHPUnit\Framework\TestCase;
  */
 abstract class UnitTestCase extends TestCase
 {
-    private bool $initialized = false;
-    private ?Connection $connection = null;
-
-    /** @after */
-    protected function disconnect(): void
-    {
-        if ($this->connection) {
-            while ($this->connection->isTransactionActive()) {
-                $this->connection->rollBack();
-            }
-            unset($this->connection);
-        }
-    }
-
     /**
-     * Create database connection.
-     *
-     * Default will use a mock object.
+     * Create new query builder.
      */
-    protected function createConnection(): Connection
+    protected function getDatabaseSession(): DatabaseSession
     {
-        $connection = $this->createMock(Connection::class);
-
-        $expressionBuilder = new ExpressionBuilder($connection);
-
-        $connection
-            ->expects(self::any())
-            ->method('getExpressionBuilder')
-            ->willReturn($expressionBuilder)
-        ;
-
-        $platform = $this->createPlatform();
-
-        $connection
-            ->expects(self::any())
-            ->method('getDatabasePlatform')
-            ->willReturn($platform)
-        ;
-
-        return $connection;
-    }
-
-    /**
-     * Create database platform accordingly to environment configuration.
-     *
-     * We need a default one so for now, it's arbitrarily PostgreSQL.
-     *
-     * This method will be ignored in functional tests.
-     */
-    protected function createPlatform(): AbstractPlatform
-    {
-        return new PostgreSQLPlatform();
-    }
-
-    /**
-     * Initialize database.
-     */
-    protected function initializeDatabase(): void {}
-
-    /**
-     * Get testing connection object.
-     */
-    protected function getConnection(): Connection
-    {
-        if (!$this->initialized) {
-            $this->initializeDatabase();
-
-            $this->initialized = true;
-        }
-
-        return $this->connection ??= $this->createConnection();
+        return new MockQueryBuilder();
     }
 
     /**