+++ tests/Functional/Anonymizer/Core/ConstantAnonymizerTest.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Tests\Functional\Anonymizer\Core;
 
-use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizationConfig;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizator;
-use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\AnonymizerRegistry;
 use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\Options;
+use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
 use MakinaCorpus\DbToolsBundle\Test\FunctionalTestCase;
 
 class ConstantAnonymizerTest extends FunctionalTestCase
@@ -23,45 +20,38 @@ class ConstantAnonymizerTest extends FunctionalTestCase
             ],
             [
                 [
-                    'id' => '1',
+                    'id' => 1,
                     'data' => 'toto1@example.com',
                 ],
                 [
-                    'id' => '2',
+                    'id' => 2,
                     'data' => 'toto2@example.com',
                 ],
                 [
-                    'id' => '3',
+                    'id' => 3,
                     'data' => 'toto3@example.com',
                 ],
                 [
-                    'id' => '4',
+                    'id' => 4,
                 ],
             ],
         );
 
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'constant',
             new Options(['value' => 'xxxxxx'])
         ));
 
-        $anonymizator = new Anonymizator(
-            $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             "toto1@example.com",
-            $this->getConnection()->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $this->getConnection()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $this->assertSame('xxxxxx', $datas[0]);
         $this->assertSame('xxxxxx', $datas[1]);
@@ -79,45 +69,38 @@ class ConstantAnonymizerTest extends FunctionalTestCase
             ],
             [
                 [
-                    'id' => '1',
-                    'data' => '52',
+                    'id' => 1,
+                    'data' => 52,
                 ],
                 [
-                    'id' => '2',
-                    'data' => '76',
+                    'id' => 2,
+                    'data' => 76,
                 ],
                 [
-                    'id' => '3',
-                    'data' => '89',
+                    'id' => 3,
+                    'data' => 89,
                 ],
                 [
-                    'id' => '4',
+                    'id' => 4,
                 ],
             ],
         );
 
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'constant',
             new Options(['value' => '2012', 'type' => 'integer'])
         ));
 
-        $anonymizator = new Anonymizator(
-            $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             52,
-            (int) $this->getConnection()->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            (int) $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $this->getConnection()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $this->assertSame(2012, (int) $datas[0]);
         $this->assertSame(2012, (int) $datas[1]);