+++ tests/Functional/Anonymizer/Core/FloatAnonymizerTest.php
@@ -4,11 +4,7 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Tests\Functional\Anonymizer\Core;
 
-use Doctrine\DBAL\Platforms\SQLServerPlatform;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizator;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\AnonymizerRegistry;
 use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\Options;
-use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizationConfig;
 use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
 use MakinaCorpus\DbToolsBundle\Test\FunctionalTestCase;
 
@@ -25,19 +21,19 @@ class FloatAnonymizerTest extends FunctionalTestCase
             ],
             [
                 [
-                    'id' => '1',
-                    'data' => '10.5',
+                    'id' => 1,
+                    'data' => 10.5,
                 ],
                 [
-                    'id' => '2',
-                    'data' => '20.5',
+                    'id' => 2,
+                    'data' => 20.5,
                 ],
                 [
-                    'id' => '3',
-                    'data' => '30.5',
+                    'id' => 3,
+                    'data' => 30.5,
                 ],
                 [
-                    'id' => '4',
+                    'id' => 4,
                 ],
             ],
         );
@@ -45,28 +41,21 @@ class FloatAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithMinAndMax(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'float',
             new Options(['min' => 2, 'max' => 5.5, 'precision' => 6])
         ));
 
-        $anonymizator = new Anonymizator(
-            $connection = $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             10.5,
-            (float) $connection->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            (float) $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $connection->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $data = (float) $datas[0];
         $this->assertNotNull($data);
@@ -99,28 +88,21 @@ class FloatAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithDelta(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'float',
             new Options(['delta' => 5.2, 'precision' => 4])
         ));
 
-        $anonymizator = new Anonymizator(
-            $connection = $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             10.5,
-            (float) $connection->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            (float) $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $connection->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $data = (float) $datas[0];
         $this->assertNotNull($data);
@@ -154,28 +136,21 @@ class FloatAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithPercent(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'float',
             new Options(['percent' => 5])
         ));
 
-        $anonymizator = new Anonymizator(
-            $connection = $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             10.5,
-            (float) $connection->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            (float) $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $connection->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $data = (float) $datas[0];
         $this->assertNotNull($data);