+++ tests/Functional/Anonymizer/Core/LoremIpsumAnonymizerTest.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Tests\Functional\Anonymizer\Core;
 
-use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizationConfig;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizator;
-use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
-use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\AnonymizerRegistry;
 use MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\Options;
+use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
 use MakinaCorpus\DbToolsBundle\Test\FunctionalTestCase;
 
 class LoremIpsumAnonymizerTest extends FunctionalTestCase
@@ -24,19 +21,19 @@ class LoremIpsumAnonymizerTest extends FunctionalTestCase
             ],
             [
                 [
-                    'id' => '1',
+                    'id' => 1,
                     'data' => 'test1',
                 ],
                 [
-                    'id' => '2',
+                    'id' => 2,
                     'data' => 'test2',
                 ],
                 [
-                    'id' => '3',
+                    'id' => 3,
                     'data' => 'test3',
                 ],
                 [
-                    'id' => '4',
+                    'id' => 4,
                 ],
             ],
         );
@@ -44,28 +41,21 @@ class LoremIpsumAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithDefaultOptions(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'lorem',
             new Options()
         ));
 
-        $anonymizator = new Anonymizator(
-            $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             'test1',
-            $this->getConnection()->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $this->getConnection()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         // Default behavior is to create on paragraph, without HTML '<p>' tag.
         $data = $datas[0];
@@ -93,8 +83,7 @@ class LoremIpsumAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithParagraphsAndTag(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'lorem',
@@ -104,20 +93,14 @@ class LoremIpsumAnonymizerTest extends FunctionalTestCase
             ])
         ));
 
-        $anonymizator = new Anonymizator(
-            $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             'test1',
-            $this->getConnection()->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $this->getConnection()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $data = $datas[0];
         $this->assertNotNull($data);
@@ -142,28 +125,21 @@ class LoremIpsumAnonymizerTest extends FunctionalTestCase
 
     public function testAnonymizeWithWords(): void
     {
-        $config = new AnonymizationConfig();
-        $config->add(new AnonymizerConfig(
+        $anonymizator = $this->createAnonymizatorWithConfig(new AnonymizerConfig(
             'table_test',
             'data',
             'lorem',
             new Options(['words' => 5])
         ));
 
-        $anonymizator = new Anonymizator(
-            $this->getConnection(),
-            new AnonymizerRegistry(),
-            $config
-        );
-
         $this->assertSame(
             'test1',
-            $this->getConnection()->executeQuery('select data from table_test where id = 1')->fetchOne(),
+            $this->getDatabaseSession()->executeQuery('select data from table_test where id = 1')->fetchOne(),
         );
 
         $anonymizator->anonymize();
 
-        $datas = $this->getConnection()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
+        $datas = $this->getDatabaseSession()->executeQuery('select data from table_test order by id asc')->fetchFirstColumn();
 
         $data = $datas[0];
         $this->assertNotNull($data);