+++ tests/Unit/Anonymization/Anonymizer/Core/FloatAnonymizerTest.php
@@ -15,7 +15,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'min' => 12.5,
                 'max' => 14.5,
@@ -32,7 +32,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'min' => 15.5,
                 'max' => 14.5,
@@ -45,7 +45,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'delta' => 12.5,
             ]),
@@ -61,7 +61,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'delta' => -12.5,
             ]),
@@ -73,7 +73,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'percent' => 12,
             ]),
@@ -89,7 +89,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'percent' => -12,
             ]),
@@ -103,7 +103,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'min' => 12.5,
                 'max' => 14.5,
@@ -116,7 +116,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'min' => 12.5,
                 'max' => 14.5,
@@ -129,7 +129,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([
                 'delta' => 14.5,
                 'percent' => 14.5,
@@ -144,7 +144,7 @@ class FloatAnonymizerTest extends UnitTestCase
         new FloatAnonymizer(
             'some_table',
             'some_column',
-            $this->getConnection(),
+            $this->getDatabaseSession(),
             new Options([]),
         );
     }